Refactor - Moving Dialog components
