+++ anubis-family/src/app/app.module.ts
@@ -15,10 +15,10 @@ import {HttpClientModule} from '@angular/common/http';
 import {DailyTaskComponent} from './component/daily-tasks/daily-task/daily-task.component';
 import {RegisterComponent} from './component/register/register.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {RegisterDialogComponent} from './component/register/register-dialog.component';
+import {RegisterDialogComponent} from './component/dialog/register-dialog/register-dialog.component';
 import {authInterceptorProviders} from './interceptor/auth.interceptor';
 import {LoginComponent} from './component/login/login.component';
-import {LoginDialogComponent} from './component/login/login-dialog/login-dialog.component';
+import {LoginDialogComponent} from './component/dialog/login-dialog/login-dialog.component';
 import {CalendarComponent} from './component/calendar/calendar.component';
 import {ProfileComponent} from './component/profile/profile.component';
 import {DailyTaskCreateComponent} from './component/daily-tasks/daily-task-create/daily-task-create.component';
