+++ anubis-family/src/app/component/dialog/register-dialog/register-dialog.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
 import {MatDialogRef} from '@angular/material/dialog';
-import {RegisterComponent} from './register.component';
+import {RegisterComponent} from '../../register/register.component';
 
 @Component({
   selector: 'app-register-dialog',
