+++ anubis-family/src/app/component/navigation/navigation.component.ts
@@ -3,8 +3,8 @@ import {BreakpointObserver, Breakpoints} from '@angular/cdk/layout';
 import {Observable} from 'rxjs';
 import {map, shareReplay} from 'rxjs/operators';
 import {MatDialog} from '@angular/material/dialog';
-import {RegisterDialogComponent} from '../register/register-dialog.component';
-import {LoginDialogComponent} from '../login/login-dialog/login-dialog.component';
+import {RegisterDialogComponent} from '../dialog/register-dialog/register-dialog.component';
+import {LoginDialogComponent} from '../dialog/login-dialog/login-dialog.component';
 import {TokenStorageService} from '../../service/token-storage.service';
 import {Router} from '@angular/router';
